pyproject.toml
[tool.poetry]
name = "papers-visualization"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "papers_visualization"}]

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
pandas = "^2.0.2"
numpy = "^1.24.3"
requests = "^2.31.0"
scikit-learn = "^1.2.2"
torch = "^2.0.1"
selenium = "^4.10.0"
tqdm = "^4.65.0"
transformers = "^4.30.2"
accelerate = "^0.20.3"
umap-learn = "^0.5.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"